+++ packages/melt/src/lib/builders/Accordion.spec.svelte.ts
@@ -1,8 +1,9 @@
 import { expectTypeOf, it } from "vitest";
 import { Accordion, type AccordionProps } from "./Accordion.svelte";
 import type { SvelteSet } from "svelte/reactivity";
+import { testWithEffect } from "$lib/utils/test.svelte";
 
-it("Should have valid types", () => {
+testWithEffect("Should have valid types", () => {
 	const a1 = new Accordion();
 	expectTypeOf(a1.value).toMatchTypeOf<string | undefined>();