+++ packages/melt/src/lib/builders/index.ts
@@ -11,3 +11,7 @@ export * from "./Toggle.svelte";
 export * from "./Slider.svelte";
 export * from "./Tree.svelte";
 export * from "./utils.svelte";
+export * from "./FileUpload.svelte";
+
+
+export * from './Select.svelte';