+++ packages/melt/CHANGELOG.md
@@ -1,5 +1,69 @@
 # melt
 
+## 0.13.0
+
+### Minor Changes
+
+- add wip select ([`6301b0c`](https://github.com/melt-ui/next-gen/commit/6301b0c494a903460f2ddbf8dd7d3966d52ca180))
+
+- tree: allow passing dynamic iterables to item ([`25f7f21`](https://github.com/melt-ui/next-gen/commit/25f7f21ad84f7171c390c4e1a2ac2156d0c3ab69))
+
+- popover: add sameWidth prop ([`99771a8`](https://github.com/melt-ui/next-gen/commit/99771a83e1b0df7ce4997e59db5067d3fc58dce4))
+
+### Patch Changes
+
+- auto add position: absolute to popover content ([`99771a8`](https://github.com/melt-ui/next-gen/commit/99771a83e1b0df7ce4997e59db5067d3fc58dce4))
+
+- type fixes ([`f77c910`](https://github.com/melt-ui/next-gen/commit/f77c9101597ec2005954586d561499c0cbb30649))
+
+## 0.12.0
+
+### Minor Changes
+
+- 6e98362: tree: allow passing dynamic iterables to item
+
+## 0.11.0
+
+### Minor Changes
+
+- c4d3f96: add file upload
+
+## 0.10.3
+
+### Patch Changes
+
+- a1fa673: fix deep merge fn
+
+## 0.10.2
+
+### Patch Changes
+
+- f7767fa: add floating ui to peer deps
+
+## 0.10.1
+
+### Patch Changes
+
+- 476006d: fix accordion component import
+
+## 0.10.0
+
+### Minor Changes
+
+- 979c893: export builder instance in components
+
+## 0.9.0
+
+### Minor Changes
+
+- 0886281: BREAKING: remove collapsible root element
+
+## 0.8.1
+
+### Patch Changes
+
+- b30421f: fix component types
+
 ## 0.8.0
 
 ### Minor Changes